+++ src/gameboard.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use std::collections::HashSet;
 
 // mod crate:quadri_enumerator;
@@ -5,14 +6,15 @@ use std::collections::HashSet;
 use crate::quadri_enumerator;
 use crate::utils;
 
-static X_DIM: usize = 4;
-static Y_DIM: usize = 4;
+pub static X_DIM: usize = 4;
+pub static Y_DIM: usize = 4;
 
 #[derive(Debug)]
 pub struct Gameboard {
     pub board: Vec<Option<usize>>, //Stored in 1 d, managed rows/columns via helpers
     pub ix_as_alpha: Vec<String>,
-    pub active_ixs: HashSet<usize>,
+    pub active_piece_ixs: HashSet<usize>,
+    pub piece_ix_to_board_ix: HashMap<usize, usize>,
     pub empty_spaces: HashSet<[usize; 2]>,
     pub used_spaces: HashSet<[usize; 2]>,
     pub quadri_coords: Vec<Vec<[usize; 2]>>,
@@ -61,7 +63,8 @@ impl Gameboard {
         Gameboard {
             board: this_board,
             ix_as_alpha: ix_as_alpha,
-            active_ixs: HashSet::new(),
+            active_piece_ixs: HashSet::new(),
+            piece_ix_to_board_ix: HashMap::new(),
             empty_spaces: empty_spaces,
             used_spaces: HashSet::new(),
             quadri_coords: quadri_coords,
@@ -92,7 +95,7 @@ impl Gameboard {
     }
     pub fn place_piece(&mut self, piece_index: usize, x: usize, y: usize) -> Result<usize, String> {
         println!("piece_index: {}, x: {}, y: {}", piece_index, x, y);
-        if let Some(_ix) = self.active_ixs.get(&piece_index) {
+        if let Some(_ix) = self.active_piece_ixs.get(&piece_index) {
             return Err("This piece already on board".to_string());
         }
         if self.used_spaces.contains(&[x, y]) {
@@ -100,7 +103,8 @@ impl Gameboard {
         }
         let board_ix = self.coord_to_ix(&x, &y);
         self.board[board_ix] = Some(piece_index);
-        self.active_ixs.insert(piece_index);
+        self.active_piece_ixs.insert(piece_index);
+        self.piece_ix_to_board_ix.insert(piece_index, board_ix);
         self.empty_spaces.remove(&[x, y]);
         self.used_spaces.insert([x, y]);
         Ok(piece_index)
@@ -112,9 +116,28 @@ impl Gameboard {
             .ok_or("Square is already piece-less".to_string())?;
         let board_ix = self.coord_to_ix(&x, &y);
         self.board[board_ix] = None;
-        self.active_ixs.remove(&current_piece_index);
+        self.active_piece_ixs.remove(&current_piece_index);
+        self.piece_ix_to_board_ix.remove(&current_piece_index);
         self.empty_spaces.insert([x, y]);
         self.used_spaces.remove(&[x, y]);
         Ok(current_piece_index)
     }
+
+    pub fn get_board_ix_for_piece_ix(&self, piece_ix: &usize) -> Option<&usize> {
+        self.piece_ix_to_board_ix.get(&piece_ix)
+    }
+
+    pub fn get_coord_for_piece_ix(&self, piece_ix: &usize) -> Option<[usize; 2]> {
+        match self.get_board_ix_for_piece_ix(&piece_ix) {
+            None => None,
+            Some(u) => Some(self.ix_to_coord(&u)),
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    // use super::*;
+
+    //TODO - write tests
 }
