+++ src/lib.rs
@@ -1,4 +1,5 @@
-pub mod gameboard;
-pub mod gamepieces;
-pub mod quadri_enumerator;
-pub mod utils;
+pub mod game;
+mod gameboard;
+mod gamepieces;
+mod quadri_enumerator;
+mod utils;
